Ticket #60: Do not use std::exception(const char*) since not C++ standard
The constructor of the std::exception class does not take any arguments
according to C++ standard. Microsoft C++ has a constructor taking
arguments but this is not standard. The std::runtime_error class has a
constructor taking arguments (char*) on both platforms, Windows and
Linux. To be portable, better use runtime_error.