Update Zipper flags (#89, #103)
Old Zipper::zipFlags enum was a mix between the behavior for the archive
when opened and compression factor. Now separated! The Append flag is no
longer "hard coded" as local variable but now passed from constructor
and method. The default behavior is the flag Overwrite for the
constructor (instead of the flag Append) and flag Append for open()
method (unchanged). Update the README in accordance.