Fix build error on windows
ioapi.h needs to be included before iowin32.h as iowin32.h (a minizip header pulled in as a submodule) references definitions in ioapi.h that are not defined in iowin32.h. Ideally this would be fixed in minizip's iowin32.h but simply including before #include-ing also fixes the issue at the zipper project level.